Add cucumber tests of the main sign-in flow
Uses a combination of OmniAuth's built in stubbing and the vcr gem
to handle external interactions. That means we only really test that
the app behaves properly when the google APIs behave as expected.
It would be good to have a full smoke test that does end to end testing
and can be run manually but that may have to wait a while.